requirements.txt
torch
transformers
datasets
sentencepiece
pandas
scikit-learn
requests
Pillow
numpy
